Move RateLimitUtils to core module
