+++ java/clients/src/main/java/sleeper/clients/teardown/ShutdownSystemProcesses.java
@@ -33,7 +33,7 @@ import java.util.function.Consumer;
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.COMPACTION_CLUSTER;
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.INGEST_CLUSTER;
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.SPLITTING_COMPACTION_CLUSTER;
-import static sleeper.util.RateLimitUtils.sleepForSustainedRatePerSecond;
+import static sleeper.core.util.RateLimitUtils.sleepForSustainedRatePerSecond;
 
 public class ShutdownSystemProcesses {
