+++ java/core/src/main/java/sleeper/core/util/RateLimitUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.util;
+package sleeper.core.util;
 
 public class RateLimitUtils {
